Add a "Send to Simple Neurite Tracer" panel to TrakEM2
The idea of this is to be able to trace selected parts of very
large volumes using Simple Neurite Tracer, where the size of
the volume is so large that only TrakEM2 to can deal with it
comfortably.

If you select a rectangular region of interest and a range of
layers in this panel, the stack that's defined can be sent to
Simple Neurite Tracer for tracing.